+++ CHANGELOG.md
@@ -5,8 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
--
+### `jsonschema-module-jakarta-validation`
+#### Added
+- populate `const`/`enum` based on `@AssertTrue`/`@AssertFalse`
 
 ## [4.35.0] - 2024-03-29
 ### `jsonschema-generator`