+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### `jsonschema-generator`
+#### Changed
+- consider `Boolean` values as valid in `const`/`enum` (i.e., no longer ignore them)
+
 ### `jsonschema-module-jakarta-validation`
 #### Added
 - populate `const`/`enum` based on `@AssertTrue`/`@AssertFalse`