+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [4.36.0] - 2024-07-20
 ### `jsonschema-generator`
 #### Added
 - new `Option.ACCEPT_SINGLE_VALUE_AS_ARRAY` to support Jackson `DeserializationFeature` of the same name, i.e., when an array type is declared, an instance of a single item should also be accepted by the schema
@@ -833,7 +833,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Indicate a number's "exclusiveMaximum" according to `@DecimalMax` or `@Negative`
 
 
-[Unreleased]: https://github.com/victools/jsonschema-generator/compare/v4.35.0...HEAD
+[4.36.0]: https://github.com/victools/jsonschema-generator/compare/v4.35.0...v4.36.0
 [4.35.0]: https://github.com/victools/jsonschema-generator/compare/v4.34.0...v4.35.0
 [4.34.0]: https://github.com/victools/jsonschema-generator/compare/v4.33.1...v4.34.0
 [4.33.1]: https://github.com/victools/jsonschema-generator/compare/v4.33.0...v4.33.1